TRS-80 Model 100 Day 1 done
reads data file from a backpack drive (or other TPDD emulator or actual floppy) as would likely be out of RAM if the file was on the machine itself.

Uses the BasicIDE for the model 100 at https://github.com/AyrA/BasicIDE

Gives the correct answer to both parts. Takes a while to run.